Refactor table and collection views with protocols to make UI updates shared
 - Created protocol SentMemeViewCell for all reusable cell views (table and collection) that have a UIImageView.
The MemeCollectionViewCell conforms as does UITableViewCell with an extension

 - Created protocol SentMemeViewControllers for all VCs in the tab VC. This protocol has a selectedItem property
so the tab VC can perform the segue to the detail VC. It will be able to guarantee the sender has a selectedItem
if it conforms to this protocol. Then it will be able to get the selected meme to pass on the detail VC. At the moment
no VCs conform to this base protocol, but if we had one without images this is what we'd use.

 - Created SentMemeWithImageViewControllers that inherits from SentMemeViewControllers so that we can
define func configureCellImage to set the image for any VC that conforms to SentMemeViewCell. Also
extended it to make a default implementation. Not yet implemented but it should check for a memedImage
in the meme and use it in the UIImageView, otherwise use the originalImage. Both the collection VC and
table VC conform to this.

 - Any class that conforms to SentMemeWithImageViewControllers should generate a cell with some
version of dequeueReusableCell that returns a cell that conforms to SentMemeViewCell, otherwise it won't
be able to call configureCellImage, and therefore doesn't need to conform to this protocol.

 - Renamed and rewired outlets on MemeCollectionViewCell. The imageView outlet shares it's name with
UITableViewCell's imageView outlet and conforms to the SentMemeViewCell. The textLabel is only for this
class and I'll add another one later. Note we're making the imageView a UIImageView? and not UIImageView!
to conform, so we will unwrap it anywhere we use it.