Refactor MemeEditorVC and MemeDetailVC to remove inheritance and use protocols
 - Removed SentMemeViewCell protocol. It was only used for the UIImageView, and the only change being done was to set the
image view based on the meme having a memedImage or not, which is now done in one line of code. Therefore we also
removed SentMemeWithImageViewControllers protocol and it's configureCellImage method

 - Removed conformance of MemeTableVC and MemeCollectionVC to SentMemeWithImageViewControllers and instead
just set conformance to SentMemeViewControllers. We set the cell's image view directly since we know that we're getting either
a UITableViewCell or MemeCollectionViewCell, both of which have an imageView property.

 - For the Editor and Detail VCs, there is now a MemeDetailAndEditor protocol that requires the UI elements for the meme
image and text fields, as well as a var to store the current meme. Two functions, one to configure the textfield and setup func
that calls configureText on both fields and sets the image is also there.

 - The configureText func has been moved to the MemeDetailAndEditor protocol extension. viewdidload and viewwillappear
for both VCs have been pared down. MemeDetailAndEditor conformance added to both VCs. Editor VC removed inheritance
from Detail VC and implemented conformance to MemeDetailAndEditor protocol.

 - Rewired a few outlets, and set some UI element properties through storyboard. Removed code that sets some UI element
statuses.